pyproject.toml
[project]
name = "hachimi-feeder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy<2",
    "opencv-python",
    "ultralytics",
    "sounddevice",
    "tflite_runtime",
    "rpi-gpio",
    "pyaudio",
]
